requirements.txt
streamlit
google-generativeai
Pillow
numpy